// src/api/basic/msg/model/extendMsgTemplateModel.ts
export interface ExtendMsgTemplatePageQuery {
  type?: string; // 消息类型
  code?: string; // 模板标识
  name?: string; // 模板名称
  title?: string; // 标题
  sign?: string; // 签名
  content?: string; // 模板内容
  script?: string; // 脚本
  param?: string; // 模板参数
  remarks?: string; // 备注
  target?: string; // 打开方式
  autoRead?: boolean; // 自动已读
  remindMode?: string; // 提醒方式
  url?: string; // 跳转地址
}

export interface ExtendMsgTemplateSaveVO {
  type?: string; // 消息类型
  code?: string; // 模板标识
  name?: string; // 模板名称
  title?: string; // 标题
  sign?: string; // 签名
  content?: string; // 模板内容
  script?: string; // 脚本
  param?: string; // 模板参数
  remarks?: string; // 备注
  target?: string; // 打开方式
  autoRead?: boolean; // 自动已读
  remindMode?: string; // 提醒方式
  url?: string; // 跳转地址
}

export interface ExtendMsgTemplateUpdateVO {
  id: string;
  type?: string; // 消息类型
  code?: string; // 模板标识
  name?: string; // 模板名称
  title?: string; // 标题
  sign?: string; // 签名
  content?: string; // 模板内容
  script?: string; // 脚本
  param?: string; // 模板参数
  remarks?: string; // 备注
  target?: string; // 打开方式
  autoRead?: boolean; // 自动已读
  remindMode?: string; // 提醒方式
  url?: string; // 跳转地址
}

export interface ExtendMsgTemplateResultVO {
  echoMap?: any;
  id?: string; // 模板ID
  createdBy?: string; // 创建人ID
  createdTime?: string; // 创建时间
  updatedBy?: string; // 最后修改人
  updatedTime?: string; // 最后修改时间
  type?: string; // 消息类型
  code?: string; // 模板标识
  name?: string; // 模板名称
  title?: string; // 标题
  sign?: string; // 签名
  content?: string; // 模板内容
  script?: string; // 脚本
  param?: string; // 模板参数
  remarks?: string; // 备注
  target?: string; // 打开方式
  autoRead?: boolean; // 自动已读
  remindMode?: string; // 提醒方式
  url?: string; // 跳转地址
}